+++ R/sensivitity-time-profiles.R
@@ -101,6 +101,7 @@
 #' )
 #'
 #' sensitivitySpiderPlot(resultsNamed)
+#' }
 
 #' }
 #' @export