+++ NEWS.md
@@ -1,3 +1,5 @@
+# esqlabsR (development version)
+
 # esqlabsR 5.4.0
 
 ## Breaking changes